SILGen: Don't rescope cases unless they actually bind variables in a specialization.
Check that a row in the clause matrix actually binds variables before rescoping it during pattern specialization, instead of always rescoping it if its constructor got specialized. We need to be able to `fallthrough` into scopes without variables, and fallthrough-ing across scopes isn't going to work. A nice side effect of this is that we can kill some useless continuation BBs; we won't use the continuation BB for a specialization scope unless variables actually get bound in that scope.

Swift SVN r6191