SIL: Assert that bridged types actually exist.
Instead of quietly not bridging. If the Clang importer and other module importers did their job correctly, there's no way the bridge types shouldn't be there at SILGen time. Rearrange the checks to always look for the native type first; if we never see a String, we don't also need to see an NSString, which could legitimately be missing if we didn't import Foundation and don't use any NSString APIs.

Swift SVN r6340