[ClangImporter] Don't assume forward declarations are never filled in.
Previously the result of a failed import was always cached, even if it
failed because a forward declaration was missing.

There is an actual problem here: NSString methods in the ObjectiveC
module (specifically, -description and +description, and later
-debugDescription) are not being imported if ObjectiveC is imported
before Foundation. However, previously this resulted in a crash.

Swift SVN r6524