Revert "Debug Info: Add support for enum-style oneof types."
We're redesigning enums, and the current implementation isn't really
handled in the module format (since it uses the presence or absence of
a source location to decide if a oneof element is "enum-like"). We'll
just drop this from the debug info for now.

This reverts r6462 / 9ed7fb3e70ff2cb9f08524699e89643386e8b6e0.

Swift SVN r6529