SIL.rst: Remove absurd grammar example.
None of the instructions I used to demonstrate the operand syntax actually use it. That's dumb.

Also refactor the grammar to fix an inaccurate use of 'sil-value'. We allow the '#n' value index even in places we don't require a type, but not for instruction or bb argument defs.

Swift SVN r6615