Better way of type checking ApplyExpr in presence of attributes.
Add the new ApplicableFunction constraint to represent loose function
type equality - one type is applicable to another, if their inputs
and outputs match. I've removed the hack on [auto_closure], so
type checking a call to a function with attributes
("auto_closure" or "noreturn") is handled in the same way now.

Disables a test case (added in r6598) that was relying on the previous
behavior. Doug has offered to fix it.

Swift SVN r6629