SILGen: Mark address-only missing returns unreachable too.
The epilog BB not taking an argument is enough by itself to determine whether we are allowed to fall into it; address-only returns do not use a BB argument to return. Break 'NeedsReturn' back out into its own bit and set it based on the Swift type passed to prepareEpilog.

Swift SVN r6802