Simplify checking of conforms-to constraints.
Teach the protocol-conformance checker to short-circuit when the type
we're checking has type variables (in which case we shouldn't do
implicit conformance checks as a fallback) and to avoid building
protocol conformances when the caller does not care.

Allow the constraint checker to pass types involving type variables to
the protocol-conformance checker, eliminating the need for
simplifyType(), type validation, and the early bail-out when type
variables are involved. 4% compile-time improvement for the standard
library in a debug build.


Swift SVN r6889