SIL Parser: parse builtin_function_ref.
In SILParser, factor out lookupTopDecl and lookupMember.
lookupTopDecl finds the top-level ValueDecl or Module given a name and
lookupMember finds the ValueDecl given a type and a member name.
Update comments for parseSILDeclRef to match SIL.rst.
We now handle the case where the first component of a dotted path is a
module.

In SILPrinter, print the full path for builtin_function_ref, also
"!" is optional when printing SILDeclRef.


Swift SVN r6926