[SIL] Issue a diagnostic on a non-exhaustive switch
The issue is truly diagnosed in SILGen and an unreachable instruction is
inserted in cases where the switch is not fully covered. However, we issue
the diagnostic later on in DataflowDiagnostics pass.

John has suggested that it's best to diagnose this after DCE because
we should not warn in cases where the switch is not fully covered,
but we can show that the non-covered cases are not reachable.

Swift SVN r6929