[SILGen] Add location info to branches generated from control flow ASTNodes.
This is what the current behavior is, it can change if we find this info
insufficient.

Branches formed from control flow due to IfStmt, ForLoop and other statements and
expressions will have the statement as the location. With one exception - if the
block ends with a return statement, we have the return statement as the location.
Currently, this preserves the location of the original return; before merging into
a single return block.

The instructions building the conditions of the control flow ASTNodes, have the
condition as the location. (I've added test for this, but this was already working
as expected.)

Swift SVN r6992