Honor [exported] or lack thereof for the imports of serialized modules.
This should complete support for Swift-side export control.

Note that a TU's imports are still all "re-exported" (i.e. returned from
getReexportedModules) whether they are marked [exported] or not. This is
a convenient trick to get lookup in the main source file to look at all
imported modules, but if we allow real multi-TU compilation we'll have to
revisit this.

Swift SVN r7052