SILGen: Actually set the new cast kind on specialized IsaPatterns.
When we have potentially overlapping 'is' patterns, such as multiple class or archetype patterns, and we specialize the decision tree on an 'is' pattern, we derive new 'is' patterns relative to the specialized type, but we weren't actually setting the correct kind for those new patterns. Oops. Caught by Ted.

Swift SVN r7180