SILGen: Choose pattern specialization branches by subsumption rather than orthogonality.
In the case of 'is' patterns involving classes or archetype, we need to include other type casts in the specialization based on whether the archetype or class could still dynamically match the specialized type. We were filtering casts based on orthogonality, which is wrong; we really need to filter based on whether the patterns have nonexhaustive overlap.

Swift SVN r7181