Cope with lvalue expressions more reasonable.
First, diagnose explicit lvalue expressions (e.g., &foo), which cannot
be used for anything other than initializing a [byref]. A Fix-It zaps
the '&'. Fixes <rdar://problem/13989094>.

Second, start to formalize the notion of a discarded expression. When
the expression is discarded, we won't do any further adjustments to
it. When the expression is not discarded, we make sure it's an
rvalue. There's more work to do here to diagnose all of the forms of
discarded expression that we want to.


Swift SVN r7210