Look up standard protocols and types in the standard library, not the TU.
However, the default literal types /are/ looked up in the current TU, so
that they can be changed. It's questionable whether or not this should
include "Slice" and "Dictionary", though.

If the module "swift" isn't available, we fall back to TU-based lookup,
which is necessary to build the standard library itself.

Full list of values now looked up in "swift":
  ArrayBound
  ArrayLiteralConvertible
  Bool
  BuiltinCharacterLiteralConvertible
  BuiltinFloatLiteralConvertible
  BuiltinIntegerLiteralConvertible
  C_ARGC
  C_ARGV
  CharacterLiteralConvertible
  DictionaryLiteralConvertible
  Enumerable
  Enumerator
  false
  FloatLiteralConvertible
  IntegerLiteralConvertible
  LogicValue
  MaxBuiltinFloatType
  MaxBuiltinIntegerType
  Optional (for T?)
  print
  Slice (for T[] but not as default array literal type)
  StringInterpolationConvertible
  StringLiteralConvertible
  true

Swift SVN r7262