IRGen: Implement switch_union dispatch for SinglePayloadUnions.
First, emit an LLVM switch over the extra tag bits, if we have them. For a zero (or missing) extra tag, switch over the payload among the extra inhabitants, defaulting to the case with payload if no extra inhabitant matches. For a non-zero extra tag, switch among the empty cases that didn't get an inhabitant.

Swift SVN r7344