IRGen: Start implementing the projection half of switch_union.
For each switch_union case that branches to a BB with a SIL argument, set up an intermediate LLVM BB into which we can emit the projection to that union case and hang the result off of the SIL argument's PHI nodes. This is a bit goofy since most switch_union destinations only have a single predecessor, but the "SILArgument == llvm::PHINode" assumption in IRGenSIL runs a bit deeper than I want to fight with. Implement projection for the easy cases of no-payload, single-case, and single-payload unions.

Swift SVN r7352