IRGen: Handle switch dispatch for multi-payload no-spare-bit unions.
If a multi-payload union has no spare bits in its payload, we just need to switch on the extra tag value to distinguish payloads. The empty cases get lumped under a minimal number of tags, so we do an additional switch on the payload for those cases.

Swift SVN r7457