Per John's advice (a few weeks ago), rework the 'assign' instruction to being a semantic assign.  The functional difference here is only for unowned pointers, where now the conversion from strong to unowned is implicit in the assign.
The logic behind this is that 'assign' is really part of SILGen, that
is only moved later to make it easier to make it be flow sensitive.
Since it is part of SILGen and will be using some of the type lowering
functionality used by SILGen, we should play by its rules.

No functionality change.


Swift SVN r7507