Fix type printing harder.
(Specifically, harder than Stephen's commit in r7510.)

The type grammar has a notion of "type-simple", which is roughly "any type
that can be put in an array without parens". Some types are normally not
printed in a type-simple-compatible way (e.g. functions "Int -> Int"), so
we fix this by adding parens when printing them as part of a larger type.

Swift SVN r7528