Introduce Type::findIf() to find a type that corresponds to some predicate.
Rather than abusing Type::transform()s short-circuit nature when
looking for types with a specific property somewhere in the type
(e.g., the presence of a generic type parameter type), introduce a
separate Type::findIf() to search for child types with some
property. As a bonus, we don't need an ASTContext for such queries.


Swift SVN r7531